Encapsulate receiving bets, sending prizes and updating last_paging_token into StellarCroupier; Move logic from Croupier.php to StellarCroupier.php; Encapsulate making a test bet into StellarCroupier.php
