+++ models/StellarCroupier.php
@@ -10,7 +10,15 @@ use ZuluCrypto\StellarSdk\Model\Payment;
 
 class StellarCroupier extends StellarServer
 {
-    public const PRIZE_MEMO_TEXT = 'x%d Winner Prize';
+    private const PRIZE_MEMO_TEXT = 'x%d Winner Prize';
+
+    public const BET_MINIMUM_AMOUNT = 0.001; // XLM
+    // minimum croupier balance that is reserved for account staying active
+    public const BALANCE_RESERVE_AMOUNT = 5; // XLM
+    // croupier balance percent reserved, so prize won't exceed (100 - self::PRIZE_RESERVE_PERCENT) % of the balance
+    public const PRIZE_RESERVE_PERCENT = 20; // %
+    public const PRIZE_RETURN_PERCENT = 95; // %
+    public const WINNER_RATES = [2, 3, 4, 5, 10, 20, 50, 100, 500, 1000, 10000, 100000, 1000000];
 
     private ?Account $croupierAccount = null;
     private float $croupierBalance;
@@ -75,6 +83,75 @@ class StellarCroupier extends StellarServer
     }
 
     /**
+     * Returns array with scheme
+     *
+     * ```php
+     * [
+     *     'bets_count' => int,
+     *     'wins' => [
+     *         [
+     *              'player_public_key' => string,
+     *              'prize_amount' => float,
+     *              'winner_rate' => int,
+     *         ]
+     *     ]
+     * ]
+     * ```
+     *
+     * @return array
+     * @throws \ErrorException
+     * @throws \ZuluCrypto\StellarSdk\Horizon\Exception\HorizonException
+     * @throws \ZuluCrypto\StellarSdk\Horizon\Exception\PostTransactionException
+     */
+    public function sendPrizesToPlayers(): array
+    {
+        $sinceCursor = StellarCroupierData::getLastPagingToken();
+
+        $bets = $this->getBets($sinceCursor);
+        $wins = [];
+        foreach ($bets as [
+                 'player_public_key' => $playerPublicKey,
+                 'bet_amount' => $betAmount,
+                 'paging_token' => $pagingToken,
+        ]) {
+            if ($result = $this->prizeAmount($betAmount)) {
+                $this->sendPrize($playerPublicKey, $result['prize_amount'], $result['winner_rate']);
+                $wins[] = [
+                    'player_public_key' => $playerPublicKey,
+                    'prize_amount' => $result['prize_amount'],
+                    'winner_rate' => $result['winner_rate'],
+                ];
+            }
+
+            StellarCroupierData::setLastPagingToken($pagingToken);
+        }
+
+        return [
+            'bets_count' => count($bets),
+            'wins' => $wins,
+        ];
+    }
+
+    /**
+     * Returns true if successfully send bet to Croupier account
+     *
+     * @param string $sourcePublicKey
+     * @param string $sourcePrivateKey
+     * @param float $amount
+     * @return bool
+     * @throws \ErrorException
+     * @throws \ZuluCrypto\StellarSdk\Horizon\Exception\PostTransactionException
+     */
+    public function makeBet(string $sourcePublicKey, string $sourcePrivateKey, float $amount): bool
+    {
+        $response = $this
+            ->buildTransaction($sourcePublicKey)
+            ->addLumenPayment(self::getCroupierPublicKey(), $amount)
+            ->submit($sourcePrivateKey);
+        return $response->getResult()->succeeded();
+    }
+
+    /**
      * @throws \ErrorException
      * @throws \ZuluCrypto\StellarSdk\Horizon\Exception\HorizonException
      */
@@ -91,7 +168,7 @@ class StellarCroupier extends StellarServer
      * @throws \ErrorException
      * @throws \ZuluCrypto\StellarSdk\Horizon\Exception\PostTransactionException
      */
-    public function sendPrize(string $destinationPublicKey, float $amount, int $winnerRate)
+    private function sendPrize(string $destinationPublicKey, float $amount, int $winnerRate)
     {
         $response = $this
             ->buildTransaction(StellarServer::getCroupierPublicKey())
@@ -99,5 +176,73 @@ class StellarCroupier extends StellarServer
             ->setTextMemo(sprintf(self::PRIZE_MEMO_TEXT, $winnerRate))
             ->submit(StellarServer::getCroupierPrivateKey());
         $this->croupierBalance -= $amount + $response->getResult()->getFeeCharged()->getScaledValue();
+    }
+
+    /**
+     * Prize amount based on bet amount, croupier balance, randomly chose win rate and randomly chose option
+     * (has won or has lost)
+     *
+     * Returns array with scheme
+     *
+     * ```php
+     * [
+     *    'winner_rate' => int,
+     *    'prize_amount' => float,
+     * ]
+     * ```
+     *
+     * If player lost, empty array returned
+     *
+     * @param float $betAmount
+     * @return array
+     * @throws \ErrorException
+     * @throws \ZuluCrypto\StellarSdk\Horizon\Exception\HorizonException
+     */
+    private function prizeAmount(float $betAmount): array
+    {
+        $croupierBalance = $this->getCroupierBalance();
+
+        if ($betAmount < self::BET_MINIMUM_AMOUNT) {
+            return [];
+        }
+
+        $winnerRate = self::WINNER_RATES[array_rand(self::WINNER_RATES)];
+        $winChance = (self::PRIZE_RETURN_PERCENT / 100) / $winnerRate;
+
+        if (!self::generateBool($winChance)) {
+            return [];
+        }
+
+        $prePrizeAmount = $betAmount * $winnerRate;
+        $prizeAmount = min($prePrizeAmount, self::maximalPrizeAmount($croupierBalance));
+
+        if ($prizeAmount < 0.000_000_1) {
+            return [];
+        }
+
+        if ($this->isTestnet()) {
+            $prizeAmount = 0.001;
+        }
+
+        return [
+            'winner_rate' => $winnerRate,
+            'prize_amount' => $prizeAmount,
+        ];
+    }
+
+    private static function generateBool(float $probability): bool
+    {
+        $rand = lcg_value(); // in range [0; 1]
+        return $rand <= $probability;
+    }
+
+    private static function maximalPrizeAmount(float $balance): float
+    {
+        $balance -= self::BALANCE_RESERVE_AMOUNT;
+        $balance *= (1 - (self::PRIZE_RESERVE_PERCENT / 100));
+        if ($balance < 0.01) {
+            $balance = 0;
+        }
+        return $balance;
     }
 }
